reorganized query handlers
