+++ src/javascript/JContent/ContentRoute/ContentLayout/queryHandlers/BaseQueryHandler.js
@@ -1,0 +1,39 @@
+import {BaseChildrenQuery} from '~/JContent/ContentRoute/ContentLayout/queryHandlers/BaseQueryHandler.gql-queries';
+
+export const BaseQueryHandler = {
+    getQuery() {
+        return BaseChildrenQuery;
+    },
+
+    getResultsPath(data) {
+        return data && data.jcr && data.jcr.nodeByPath && data.jcr.nodeByPath.children;
+    },
+
+    getFragments() {
+        return [];
+    },
+
+    getQueryParams({path, lang, uilang, pagination, typeFilter, recursionTypesFilter, sort}) {
+        return {
+            path: path,
+            language: lang,
+            displayLanguage: uilang,
+            offset: pagination.currentPage * pagination.pageSize,
+            limit: pagination.pageSize,
+            typeFilter: typeFilter,
+            fieldSorter: sort.orderBy === '' ? null : {
+                sortType: sort.order === '' ? null : (sort.order === 'DESC' ? 'ASC' : 'DESC'),
+                fieldName: sort.orderBy === '' ? null : sort.orderBy,
+                ignoreCase: true
+            },
+            ...(recursionTypesFilter ? {
+                recursionTypesFilter: recursionTypesFilter,
+                fieldGrouping: {
+                    fieldName: 'primaryNodeType.name',
+                    groups: recursionTypesFilter,
+                    groupingType: 'START'
+                }
+            } : {})
+        };
+    }
+};
