+++ src/javascript/JContent/ContentRoute/ContentLayout/queryHandlers/FilesQueryHandler.gql-queries.js
@@ -1,0 +1,27 @@
+import gql from 'graphql-tag';
+
+export const imageFields = {
+    gql: gql`
+        fragment ImageNodeFields on JCRNode {
+            width: property(name: "j:width") {
+                value
+            }
+            height: property(name: "j:height") {
+                value
+            }
+            children(typesFilter: {types: ["jnt:resource"]}) {
+                nodes {
+                    ...NodeCacheRequiredFields
+                    data: property(name: "jcr:data") {
+                        size
+                    }
+                    mimeType: property(name: "jcr:mimeType") {
+                        value
+                    }
+                }
+            }
+        }
+    `,
+    applyFor: 'node'
+};
+
